requirements.txt
mcp
tavily-python
python-dotenv
